wqe/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
elasticsearch==8.11.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
python-dateutil==2.8.2
pytz==2023.3.post1
transformers==4.34.0
torch==2.1.0
streamlit==1.28.1
plotly==5.18.0
pandas==2.1.3
openai==1.3.0
numpy==1.24.3